+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 * WaitAndTap, waitAndCheck and waitAndType methods
 * OverridePermissions method
 * SetGeolocation and geolocation methods
+* Puppeteer typings updated
+* Minor improvement and fix in linting process
 
 2.8.0 / 2019-11-07
 ==================